updating docker-compose and renaming contracts
