+++ scripts/emulations/contracts/cere01-metadata.json
