+++ scripts/emulations/contracts/cere01.wasm
