+++ scripts/emulations/contracts/cere02-metadata.json
