+++ scripts/emulations/contracts/cere02.wasm
