+++ scripts/emulations/ddc-smart-contract.ts
@@ -6,7 +6,7 @@ import {
   CodePromise,
   ContractPromise,
 } from "@polkadot/api-contract";
-import cere02Abi from "./contract/cere02-metadata.json";
+import cere02Abi from "./contracts/cere02-metadata.json";
 import fs from "fs";
 const cere02Wasm = fs.readFileSync("./contract/cere02.wasm");
 import EventHandlers from "./event-handlers";
